# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Citations API",
        "version": "1"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authorization",
                "description": "Authorization and return API-token",
                "operationId": "13b5b42fe243c504cbcb70a25a54bc1e",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-mail for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User and API-token"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registration of new user",
                "description": "Registration of new user and return API-token",
                "operationId": "31416ef0da616084de2ce2b466b9ce27",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "E-mail for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User and API-token"
                    }
                }
            }
        },
        "/citations": {
            "get": {
                "tags": [
                    "Citations"
                ],
                "summary": "List of citations",
                "description": "List of citations",
                "operationId": "0772f9bdd202f3088adec2f7bb773618",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Coint of citations in list",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The list of citations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Citations"
                ],
                "summary": "Update citation",
                "description": "Update citation",
                "operationId": "f55b37404d7abc0e950c93bfc6a9f039",
                "parameters": [
                    {
                        "name": "citation_id",
                        "in": "query",
                        "description": "Id of citation in table `citations`",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "citation",
                        "in": "query",
                        "description": "Text of citation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The list of citations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}